+++ resources/views/livewire/admin/members.blade.php
@@ -1,8 +1,10 @@
 <div class="p-6">
     <div class="flex items-center justify-end px-4 py-3 text-right sm:px-6">
-        <x-jet-button wire:click="createShowModal">
-            {{ __('Create') }}
-        </x-jet-button>
+        @can('create', App\Models\Member::class)
+            <x-jet-button wire:click="createShowModal">
+                {{ __('Create') }}
+            </x-jet-button>
+        @endcan
     </div>
 
     {{-- The data table --}}
@@ -29,12 +31,16 @@
                                         <td class="px-6 py-2">{{ $item->phone }}</td>
                                         <td class="px-6 py-2">{{ $item->mobile }}</td>
                                         <td class="px-6 py-2 flex justify-end">
-                                            <x-jet-update-button wire:click="updateShowModal({{ $item->id }})">
-                                                {{ __('Update') }}
-                                            </x-jet-button>
-                                            <x-jet-danger-button class="ml-2" wire:click="deleteShowModal({{ $item->id }})">
-                                                {{ __('Delete') }}
-                                            </x-jet-button>
+                                            @can('update', App\Models\Member::class)
+                                                <x-jet-update-button wire:click="updateShowModal({{ $item->id }})">
+                                                    {{ __('Update') }}
+                                                </x-jet-button>
+                                            @endcan
+                                            @can('delete', App\Models\Member::class)
+                                                <x-jet-danger-button class="ml-2" wire:click="deleteShowModal({{ $item->id }})">
+                                                    {{ __('Delete') }}
+                                                </x-jet-button>
+                                            @endcan
                                         </td>
                                     </tr>
                                 @endforeach
